requirements.txt
amqp==5.0.6
APScheduler==3.7.0
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
dateutils==0.6.7
defusedxml==0.7.1
diff-match-patch==20200713
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.6
django-admin-rangefilter==0.8.1
django-apscheduler==0.6.0
django-background-tasks==1.2.5
django-common-helpers==0.9.2
django-compat==1.0.15
django-cors-headers==3.8.0
django-crequest==2018.5.11
django-crispy-forms==1.8.0
django-cron==0.5.1
django-crontab==0.6.0
django-formtools==2.3
django-import-export==2.5.0
django-jazzmin==2.4.8
django-polymorphic==3.0.0
django-rest-passwordreset==1.2.0
django-reversion==4.0.0
django-slick-reporting==0.6.0
django-tabular-permissions==2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
dnspython==2.1.0
et-xmlfile==1.1.0
eventlet==0.32.0
filelock==3.0.12
geographiclib==1.52
geopy==2.2.0
greenlet==1.1.1
# gunicorn==20.1.0
# heroku==0.1.4
# idna==3.2
kombu==5.1.0
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.7
Pillow==8.3.1
platformdirs==2.2.0
portalocker==2.3.2
prompt-toolkit==3.0.20
psycopg2==2.9.1
psycopg2-binary==2.9.1
PyJWT==2.1.0
# pytz==2021.1
# pywin32==301
PyYAML==5.4.1
qrcode==7.3
redis==3.5.3
requests==2.26.0
simplejson==3.17.5
six==1.16.0
soupsieve==2.2.1
sqlparse==0.4.1
swapper==1.1.1
tablib==3.0.0
tzlocal==2.1
unicodecsv==0.14.1
urllib3==1.26.7
vine==5.0.0
virtualenv==20.7.2
wcwidth==0.2.5
xlrd==2.0.1
xlwt==1.3.0
mysqlclient==2.1.0
gunicorn==19.0.0
django-debug-toolbar==3.2.4
qrcode==7.3
